+++ packages/widgetbook/lib/src/knobs/knobs.dart
@@ -1,8 +1,10 @@
 export 'boolean_knob.dart';
+export 'builders/double_knobs_builder.dart';
+export 'builders/knobs_builder.dart';
+export 'builders/knobs_extension.dart';
 export 'color_knob.dart';
 export 'double_input_knob.dart';
 export 'double_slider_knob.dart';
 export 'knob.dart';
-export 'knobs_builder.dart';
 export 'list_knob.dart';
 export 'string_knob.dart';
